pyproject.toml
[tool.poetry]
name = "dpa-dashboard"
version = "0.1.0"
description = "Developer performance analysis tool"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygithub = "^2.4.0"
plotly = "^5.24.0"
pandas = "1.5.3"
streamlit = "^1.38.0"
python-dotenv = "^1.0.1"
pandasai = "^2.2.14"
langchain = "^0.2.16"
streamlit-chat = "^0.1.1"
faiss-cpu = "^1.8.0.post1"
sentence-transformers = "^3.1.0"
langchain-community = "^0.2.16"
langchain-huggingface = "^0.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
